Claude/great allen 51nxz7

+++ src/evals/metrics/utils.py
@@ -95,8 +95,11 @@ def evaluate_probability(model, batch):
     avg_losses = losses / num_token_gt
     normalized_probs = torch.exp(-avg_losses)
 
-    avg_losses = avg_losses.cpu().numpy().tolist()
-    normalized_probs = normalized_probs.cpu().numpy().tolist()
+    # .float() is required: numpy has no bfloat16, so converting a bf16 tensor
+    # (when the model runs in bfloat16) directly via .numpy() raises
+    # "TypeError: Got unsupported ScalarType BFloat16".
+    avg_losses = avg_losses.float().cpu().numpy().tolist()
+    normalized_probs = normalized_probs.float().cpu().numpy().tolist()
     return [
         {"prob": prob, "avg_loss": avg_loss}
         for prob, avg_loss in zip(normalized_probs, avg_losses)
